Fix up comments from gemini
